change module strategy to benchmarks and sub benchmark
